型の変更
gyroの値をfroatからdoubleに

+++ main_2/gyro_serial.h
@@ -37,7 +37,7 @@ class gyroSerial {
       return;
     }
 
-    float getData() {
+    double getData() {
       return i;
     }
